fix: fixing use client error while retrieving user total spent
